requirements.txt
numpy
opencv-python
mediapipe
Pillow
